+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [Unreleased]
+## [1.2.0] - 2021-08-06
 ### Added
 - Add `use` for `EnvelopeTemplateEndpoint`
 - Add `AccountSettings.bankIdProduct`